requirements.txt
pandas
numpy
matplotlib
requests
python-crontab